AdminController: fix viewAccess template rendering
When employee does not have access to some controller error message
should be displayed.

If controller specifies custom $template in constructor, or anywhere
else in the pipeline, this error message might not be displayed
correctly. Also, custom template might not be found.

To fix this, we should always display default 'content.tpl' template
when employee does not have view access.