Remove Rijndael encryption
This encryption was originally based on deprecated mcrypt library that
is no longer available in new versions of PHP. Alternative
implementation used openssl -- the same library that is used by
PhpEncryption algorithm.

There's no need to have two encryption algorithm based on same library.
The support for Rijndael is removed.