Module: don't create is_configurable property
Code that loaded list of modules also populated dynamic property
is_configurable. This property was needed, and existed only, in this
particular code flow. For example, this property was not present when
module was instantiated directly.

To prevent PHP8.2 warnings, this commit removes creation of dynamic
property, and replaces it with method.