Modules: Remove dead code
This dead code was discovered because PHP8.2 raised warnings about
dynamic property AdminModulesController::employee.

The missing $employee property was used to query data for a view that was
no longer in use. That was the reason nobody ever reported this issue.

Fix is to remove this functionality completely. There is no reason to
run sql queries and postprocess data when we are not displaying them at
all.