Module::displayError - do not set $error property
Method Module::displayError set property Module::$error to true. Because
this property is not declared, the assignment raises warning in PHP8.2

To fix this, we could either declare property in Module class, or omit
the assignment. We decided for the later. There are more reasons.

First of ally, this property is not used anywhere in the core. There's
no need for its existence.

Also, declaring this property could be dangerous, because a lot of
modules also declared it for their own purposes. There could be
inheritance issues, for example access level mismatch. Or declared type
type could be different.

Note that there are already modules, such as native Blocknewsletter, that
declared 'error' property as private. Calling displayError() method on
such modules used to throw exception:

Cannot access private property Blocknewsletter::$error