Link::getAdminLink - check type of $params param
Thirty bees 1.4.0 added new parameter $params to the method
Link::getAdminLink() with default value.

While this should be backwards compatible change, it still causes
problems.

Some module called Link::getAdminLink() with third paramter already.
This parameter used to be ignored, but now it can cause type errors.

This commit adds type check -- if the $params parameter is not array, we
raise warning and initialize the parameter to empty array.

In the future, this check will be replaced with an actual error message
(by adding array to method signature)