Introduce method Product::isFrontOfficeContext
This commit extracts code to check if we are in front office context
to method Product::isFrontOfficeContext().

Original implementation expected that controller is always set. This
might not be always be the case, though. For example controller is not
set for CLI executions or inside work queue task. This commit fixes this
problem as well.