Update test environment
Previously, tests run in US country context. In this context a lot of
functionality, such as VAT or multilang, was not used.

To reflect that thirty bees is used mostly by EU companies, we have
switched default test environment to FR with EN language.

Some test had to be updated to pass in this new environment.