Smarty: refactor cache resource loading
Class Smarty_CacheResource_Mysql was always an odd duck. It was not
loaded by autoloader, it was not possible to override it, etc. All this
because we depended on smarty autowiring name conventions.

This commit changes this behavior. Class was renamed to match thirty
bees conventions. This new class is explicitly registered as custom
cache resource.

This is backwards compatibility change, however the impact should be
minimal.