AdminTab: display deprecation warnings
AdminTab is very old alternative to AdminController. It's quite buggy
and not maintained propery. Only few modules originally developed for
prestashop 1.4 still uses this approach.

We will remove support for this functionality in thirty bees 1.6
Since thirty bees 1.5 we will display deprecation warning, which will
hopefully prompt merchants to update the old module, or replace it.