Fix getTaxesAmount method
When parameter $priceAfterTax was passed, thirty bees tries to apply
Tools::spreadAmount function on array of floats. That raised exception,
becuase that method expects array of arrays, not arrays of scalar
values.

In lower versions of PHP this only raised warning, and didn't alter
input array in any way. In new versions of PHP, this raises error.

because calling Tools::spreadAmount didn't changed $amount array in any
way, the easiest fix is to remove this flawed functionality completely.

Closes #1628