Contact form: option to disable confirmation email
When customer submits message using contact form, thirty bees will sends
them confirmation email that includes message.

This functionality can be abused for spamming purposes. While captcha
or similar mechanisms can be used to mitigate this problem, it might not
be enough.

This feature allows merchants to decide if the confirmation email should
be send to customer in the first place.

Closes #1652