AdminOrders: do not include google maps
Core should not have any dependency on third party resources, in this
case on google maps address API.

This functionality send customer private data (address) to third party
service (google).

New hook 'displayAdminOrderAddressExtra' is called instead of displaying
google maps. This allows merchants to implement this, or similar,
functionality in a module.

Related to #1521
Related to #1278