File manager - upload files - check MIME type
This commit checks that MIME type of uploaded file matches one of the
allowed MIME types, and that file extension corresponds with it.

This posed security risks. Attacker could upload files with allowed
extension (for example image.jpg), but with dangerous content. For
example javascript of phar file. This is dangerous in combinatino with
other attacks.