Deprecate global function priceval
This function is used only for casting form inputs.

New method for all entry form cast was introduced. AdminController /
FormHelper calls this method instead of invoking 'cast' callables
directly.

This allow us to redefine build-in cast functions (for example 'intval')
and also to define custom cast functions that does not have to exists in
global namespace (such as 'priceval').