Db::getArray - throw for non-select queries
Method Db::getArray is a more strict replacement for Db::executeS. It
is possible to use Db::executeS for update queries (waring would be
displayed, though).

Becase method Db::getArray is new, and is used by thirty bees core only,
we can afford to be more strict here, and not allow update queries as
all.