Javascript CCC - fix bug with single line comment
When CCC combines file contents it simply concatenated them together. If
the input content ended with single-line comment, the result would be
incorrect javascript file.

This could happen, for example, if minification module was not
installed, or if the minifaction do not remove trailing single line
comments.

The fix is to add new lines when concatenating contents.