DB configuration: allow multi statements queries
This commit allows setting PDO option MYSQL_ATTR_MULTI_STATEMENTS

When this option is enabled, it is possible to execute multiple database
statements at once, for example

'select * from tb_customer;drop table tb_customer'

For security reasons, this option should be set to false to mitigate
impact of SQL injection attacks.

Note that this does not prevent the SQL injection attacks, or data leaks
in general. If there is SQL injection vulnerability in system, it can
still be used to access and/or modify data.