Resolve products even for wrong urls
When rewrite rule does not contain {ID}, method productID is used to
resolve it.

This method checked full product URL, and returned product ID only if
url matched. This is not really needed, we should resolve product even
with misspeled parts.

For example, if url schema is {categories:/}{rewrite}

then we should resolve product for both

https://domain.com/en/a/b/c/soap

and

https://domain.com/en/gifts/soap

If multiple products exits with the same rewrite, we still should do
full URL checks, and return the product that matches. If none of them
matches, we should resolve to the first one. It's better than displaying
Page Not Found.