Fix category reassignment in multi-store context
When category parent is changed, its position must be recalculated. In
multistore context, this needs to be performed for all stores this
category is associated with.

The map of associated shops is one-dimensional array. Code treated this
as a two-dimensinal array. This resulted in incorrect category
position, and also raised PHP warning.