List Fields: support 'callback_export' option
List fields can have 'callback' option that is called during list
rendering / export. It is usually used to add HTML code into list
output.

When exporting to CSV, such HTML code is usually not appreciated.

This commit introduces new list field option 'callback_export'. When
present, this callback will be used instead of 'callback' for export
purposes.

if 'callback_export' is not a valid callback, the original value is not
processed, and 'callback' IS NOT called. This can be used to bypass
'callback' without the need to provide specific implementation.

Example: bypass calling 'renderProductLink' method in export mode:

  [
    'title' => $this->l('Product'),
    'callback' => 'renderProductLink'
    'callback_export' => false
  ]

Example 2: call method 'renderProductLink' in export mode:

  [
    'title' => $this->l('Product'),
    'callback' => 'renderProductLink'
    'callback_export' => 'renderProductName'
  ]

Related to #1666