Multistore: Fix deleting default attribute
After method deleteDefaultAttributes is called, there must be no entry
in PREFIX_product_attribute with default_on = 1 (for this product).

In multistore environment this was not always the case. Method
updateMultishopTable() updated data in base table only if multistore
entry for current shop exists (in table PREFIX_product_attribute_shop).

Related to #1701