Product list - left join shop table
If the product id_shop_default contains shop that this product is not
associated with, the inner join makes the product disappear from the
list.

This is data corruption - this situation should never happen. Product
id_shop_default should always point to one of the shops the product is
assigned with. But there are some (not yet discovered) steps that lead
to invalid state. We have to fix that.

Meanwhile, by using left join, the product will be displayed, although
the information about the product like name and price might not be
correct.  That is because these informations are stored inside
tb_product_shop and tb_product_lang tables.

Left join also helps with situation when product is not associated with
any shops. In that case, row in tb_product exists, but there are no rows
in tb_product_shop for the product.

Related to #1769