Display carrier errors only if delivery is blocked
Method getDeliveryAddressesWithoutCarriers returns list of addresses
that we can't deliver to. This method accepts parameter $errors, that
contains information why individual carriers were rejected.

This $error array can contains informations even if method returns empty
array - this means that we can deliver to all addresses, but some
carriers were rejected for some reasons.

If we can deliver to all addresses, we don't want to display any errors
to end user. We need to display errors only when delivery is not
possible.