Attempt on fixing a bug with 0.00 carts when products are deleted
When viewing old carts where products are no longer present in the catalog current function displays 0.00 (for completed purchases, carts with green background).

I believe we should fallback to pulling the totals from the associated order (minus shipping).

This bug is also present in Cart summary (when clicking on said cart) - the list is empty and the total is again 0.00 but I was unable to make something that is working there.