Extend ShopMaintenance to delete old lines in /error_log and /admin/error_log files
Reused getLogsRetentionPeriod constant as it basically applies to those logs too.

Delete old lines and leaves those after the cutoff date.

Line by line approach for old uncleaned files that could be MB in size - should save RAM on first run.

This method appears to work with errors that don't start with a timestamp (Mail queue module) as it checks the next line for retention and removes anything above it. So even if the file is manually edited this approach should handle it.

Restores last modified date for easier debugging when looking on the date.