Fix error with cheapest product cart rules
When cart rule that applies discount to cheapest product is used,
thirty bees generates a new, single-use version of this cart rule, and
uses this one instead of the original cart rule.

We need to make sure the new generated voucher is loaded in specific
language context, otherwise exception is thrown later.

Closes #1823