Error page should not throw another error
In some rare cases, error can occur in very early stages of application
boostrap, when database or encryptor might not be initialized yet. If
that happen, error page handler would throw another exception that would
result with blank rror page.

This commit makes sure that this not happen. If encrypted message, or
email can't be resolved, we can still display error page.

Closes #1785