Move seo tags to the front of header
SEO tags (canonical, hreflangs,...) were appended to the result of
displayHeader hook. This hook can return a lot of output, pushing the
very important seo tags down the page. This can be a big problem,
because some bots do not process entire pages, but instead consider the
first n-lines only.