Fix multilang KPI refresh
Some KPI are language specific -- the precalculated value contains
translated string, for example '$100 tax excl'. These values have to be
calculated for every language individually. To achieve this, language id
has to be provided in both updateValue() and get() methods

1) ConfigurationKPI::updateValue($key, [ $languageId => $value ])'
2) ConfigurationKPI::get($key, $languageId)

For some configuration key, this was not the case. Either values update,
or values get, was not multilang. This resulted in invalid values being
displayed, and with refresh issues.

Related to #1903