Add language: ignore configuration_kpi_lang table
When new language is added to the system, thirty bees creates a new
entries for this language in all *_lang tables. We don't want this to
happen for tb_configuration_kpi_lang.

There was already an exception for tb_configuration_lang table.

Also, added an initialization callback to clean data.