getOrderByCartId - don't use shop restriction
In the Order::getOrderByCartId() method, the use of
Shop::addSqlRestriction() is unnecessary and can cause incorrect results
when not in the appropriate context. This function should return the
correct order ID when the cart ID is provided, regardless of the shop
context.

Also, added 'ORDER BY' clause to SQL statement to make this method
deterministic (in case multiple orders were created using the same cart)

Closes #1905