Fix check for specific cheapest product cart rule
When cart rule that applies to 'Cheapest product from selection' is
used, thirty bees will determine cheapest product at the time of order
creation.

Because order can be later modified in back office (the chapest product
can change), the selected cheapest product must be somehow memorized.

Somebody decided that it is a good idea to create a new, single use cart
rule to remember this information. New cart rule is generated during
checkout. This rule is a standard discount that applies to entire order,
and has monetary value that matches the value of the cheapest product.
Also, information about cheapest product is remembered in cart rule
'description' field (as a serialized json object).

When generating product tax breakdown we need cheapest product id
information.

The code simply attempted to json decode 'description' field, and then
used the result:

    if ($cheapestProduct = json_decode($cartRule->description)) {
    }

This check worked when the description contained regular text, or when
was empty. In that case, json_decode returned false.

However, this check failed when the description contained a simple
number. In that case, json_decode("123") === (int)123

This caused warning, and potentially wrong tax breakdown.

Closes #1916