Pass empty array as &$errors parameter
Method Carrier::getAvailableCarrierList allows &$errors parameter to
have null value. In some old versions of prestashop this was not the
case -- the paramter was expected to be array. There exists overrides
for this method that depend on this behaviour.

There is no reason why not to pass empty array instead of null.