Admin themes: ability to uninstall themes
It was not possible to uninstalle the theme once it was installed into
the shop. The only available action was to delete the theme (which
includes all files)

This commit adds 'Uninstall theme' option -- the theme record is removed
from database, but it remains on filesystem, and can be installed again.