List filters
This feature introduces extra filters for all back office lists. It
allows you to

- create filters on fields that are not visible in the list itself (not
  a column)
- create multiple filters on the same field
- use more advanced filter options - choose filter operator (equals,
  starts with, ends with)

We have also refactored the way filters and other list options are
persisted. We still save them to cookies, but we now have an option to
implement different storage. There are plans to do that in the future