Partial refunds: fix restock products logic
When refunding products, merchant has an option to choose if the
products should be restocked or not. This option was ignored when order
has NOT been marked as delivered. In that case, refunded products were
always re-stocked.

Since there is an UI element to govern this re-stock process, thirty
bees should not do this kind of auto-decisions on behalf of merchants.
We should always respect the selected option.

See this forum thread for more info:
https://forum.thirtybees.com/topic/6932-need-help-with-partial-refund-issue-and-reinjecting-inventory/