Show name of shop in order view if Multistore is enabled
When in order view we can't see anywhere the shop name when multistore is enabled.

And if we need to call the customer and start the conversation with 'Hello, we're calling from XYZ store' we have to go to the order list, get the shop name, get in the order again, etc.

I think it's a nice use of this free space and its not shown if Multistore is not used.