Upgrade Cart.php getProducts function
The getProducts function returns a list of products with parameters. Some of these parameters are not used anywhere at all and are actually useless.

This modification changes several of these parameters so that a promising practical opportunity to use them in templates appears. For example, in the Ajax cart you can now display discounts (-15% or -26$) and show messages to the users like "add to cart N qty to cart to apply discount."