Introduce Order::canEditProducts
This new method is used by back office order controller to decide if
product list editation is allowed or not.

Previously, this was strictly related to the delivery status. If order
went through a status with delivery flag, product list editation was
disabled.

This commit changes this by calling new method canEditProducts. The new
method still check the delivery status, for backwards compatibility
reasons. But it also calls hooks 'actionCanEditOrderProducts' that gives
modules an option to block product list editation.