Fix array to string conversion warning
If product uses ASM, thirty bees checks that product/combinations are
associated with at least one warehouses.

This check raised PHP warning because array_unique was used on array of
array. This commit simplifies this check mechanism - we no longer build
array of all warehouses, but simply fail-fast when any combination is
not associated with warehouse.

Closes #2010