Reordering customized products raised exception
Method Cart::duplicate raised an exception when it contained customized
products.

The problem is actually in Db::insert method. It tries to determine if
input $data parameter is an array of arrays, or single item (simple
array). There is a special case for arrays of (array with 'type'
key) that is used for stuff like:

	$conn->insert('table', [
	    'quantity' => [
		'type' => 'sql',
		'value' => '`quantity` '.$qty
	    ],
	]);

The parameter is array of array, but because the inner array contains
'type' key, Db class correctly treats this as as single insert.

Unfortunately, in case of Cart::duplicate, this logic causes problems.
We ultimately need to solve this problem in Db class, but it's hard to
figure out the proper and safe approach.

Until that is done, the simple fix for this instance is to call insert
method multiple times;

Closes #2016