List changed js files in AdminInformationController
Related to: https://github.com/thirtybees/thirtybees/issues/1308

After adding the js folder we should update cofing/json/filex.json by running /tools//tools/generatemd5list.php

This is done during build but I'm not sure if those on edge will experience the change until they update to official the next release.

One can run the file in FO - it then includes all js files in the list, otherwise they are shown as obsolete (which of course is wrong and they should no be removed by the merchant).

After that changes in the .js files is detected by the controller.